pyproject.toml
[tool.poetry]
name = "shellcursive"
version = "0.1.0"
description = "Recursively apply commands to any matching folders in a given path."
authors = ["Lef Filippakis <Lef-F@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.29"
rich = "^12.6.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pre-commit = "^2.21.0"
ruff = "^0.0.201"

[tool.poetry.scripts]
shellcursive = "shellcursive.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
